+++ medic-src/src/lib.rs
@@ -5,6 +5,7 @@ pub mod app_result;
 pub mod audit_step;
 pub mod check;
 pub mod config;
+pub mod doctor_step;
 pub mod runnable;
 pub mod shipit_step;
 pub mod step;
@@ -12,6 +13,7 @@ pub mod step;
 pub use app_result::AppResult;
 pub use audit_step::AuditStep;
 pub use check::Check;
+pub use doctor_step::DoctorStep;
 pub use shipit_step::ShipitStep;
 pub use step::Step;